requirements.txt
fastapi
python-docx
shortuuid
